Update firewall rules for 12000XP dongle.
The dongle that comes with 12000XP appears to be a newer revision
which does a DNS resolution for "us2.solarcloudsystem.com" instead of
using a hard-coded IP address.  Thus we update our firewall rules
to use the common port number regardless of resolved IP.

Additionally, the new dongle is probed for a "version" parameter by
the cloud, which we need to allow to let reports flow freely.